// src/features/profile/types.ts
export type PublicProfile = {
  id: string;
  fullName: string;
  email: string | null;
  phone: string | null;
  photoUrl: string | null;
  designation: string | null;
  company: string | null;
  companyFoundedYear: number | null;
  industry: string | null;
  city: string | null;
  about: string | null;
  aboutCompany: string | null;
  website: string | null;
  linkedin: string | null;
};